# ENH: Implement ppf and isf for invgauss