--- a/scipy/stats/_continuous_distns.py
+++ b/scipy/stats/_continuous_distns.py
@@ -4214,7 +4214,7 @@ class invgauss_gen(rv_continuous):
         # 1000 seems like a reasonable max number of iterations given that
         # convergence is guaranteed in this setup.
         while not all(has_converged):
-            not_done = x0[~has_converged]
+            not_done = np.nonzero(~has_converged)[0]
             lx = logF(x0[not_done], mu[not_done])
             # to avoid subtractive cancellation when `d` is very small we use a
             # taylor series expansion to approximate (p - p(xn)) as shown in
@@ -4222,11 +4222,15 @@ class invgauss_gen(rv_continuous):
             eps = logprob[not_done] - lx
             mask = np.abs(eps) < 1e-05
             delta[not_done][mask] = (
-                eps[mask] * np.exp(logprob[not_done][mask] + np.log1p(-eps[mask] / 2))
+                eps[mask] * np.exp(logprob[not_done][mask]
+                + np.log1p(-eps[mask] / 2))
             )
             delta[not_done][~mask] = q[not_done] - np.exp(lx[~mask])
 
-            x[not_done] = x0[not_done] + sign * delta[not_done] / self._pdf(x0[not_done], m[not_done])
+            x[not_done] = (
+                x0[not_done] + sign * delta[not_done]
+                / self._pdf(x0[not_done], m[not_done])
+            )
             has_converged = np.isclose(x, x0, atol=1e-08, equal_nan=True)
             x0[...] = x[...]