#### Reference issue
Follow up to #13616 and closes #13666

#### What does this implement/fix?
This PR adds methods `ppf` and `isf` to `invgauss` distribution.

#### Additional information
This PR adds an implementation of `ppf` and `isf` for `invgauss` and while also improving the `pdf` method so that it is able to return the correct value when the mean is infinity. The implementation is described in [1]. The tests are based on the examples presented in the article.

References
----------
[1] Giner, Goknur and G. Smyth. “statmod: Probability Calculations for the Inverse Gaussian Distribution.” R J. 8 (2016): 339.