@zoj613 2021-03-09
Speed of this PR compared to base branch:
```ipynb
In [4]: %timeit invgauss.ppf(p, mu=1)
15.7 ms ± 43.6 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [5]: %timeit invgauss.isf(p, mu=1)
14.2 ms ± 99.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```