@zoj613 2021-03-09
@WarrenWeckesser I believe you are working on a similar PR for norminvgauss? I'd appreciate a review/comments if you have time.