@mdhaber 2021-03-10
`norm` and `cauchy` don't seem to produce warnings with `nan` input, so it might be best to do the comparison `with np.errstate(invalid='ignore')`.