@zoj613 2021-03-10
> `norm` and `cauchy` don't seem to produce warnings with `nan` input, so it might be better to do the comparison `with np.errstate(invalid='ignore')` so to avoid the warning being raised.

are you referring to them not raising the warning in the build environment? I did not see a tests with nan input in there for cauchy. Note that there is no runtime warning raised when calling the method in a python session locally, only in the build environment.

Also: is this valid?
```
    raise AssertionError(
E   AssertionError: warning ignore filter should not be used, instead, use
E   numpy.testing.suppress_warnings (in tests only);
```

I think I remember reading somewhere that numpy's suppress_warnings shouldn't be used for new code