@zoj613 2021-03-14
the circleCI failure is not related to this PR. Seems like an issue with pythran not being installed in the build environment.