@zoj613 2021-04-09
> That one is intermittent, so I re-ran it for you. Fingers crossed.
> 
> _Update: tests passing. I think that will encourage another maintainer to take a look. I will pass on this one as I reviewed/merged [gh-13616](https://github.com/scipy/scipy/pull/13616) independently, and it would be good for another stats maintainer to become familiar with these changes before the next release._

Thanks for the update. In that case I might as well ping @andyfaff and @ev-br for a possible review.