@tupui 2021-04-10
> ModuleNotFoundError: No module named 'scipy.stats._qmc_cy'
> 
> Any idea how I can fix this @tupui ?

You might just have to merge master here. I am on master and I can build.