@zoj613 2021-04-10
> > ModuleNotFoundError: No module named 'scipy.stats._qmc_cy'
> > Any idea how I can fix this @tupui ?
> 
> You might just have to merge master here. I am on master and I can build.

Maybe im missing a step here because even after rebasing on the main branch + clearing build + re-building I get this error when importing `invgauss` locally.
```python
----> 1 from scipy.stats import invgauss

~/scipy/scipy/stats/__init__.py in <module>
    441 from .kde import gaussian_kde
    442 from . import mstats
--> 443 from . import qmc
    444 from ._multivariate import *
    445 from . import contingency

~/scipy/scipy/stats/qmc.py in <module>
    232 
    233 """
--> 234 from ._qmc import *

~/scipy/scipy/stats/_qmc.py in <module>
     14     _categorize, initialize_direction_numbers, _MAXDIM, _MAXBIT
     15 )
---> 16 from scipy.stats._qmc_cy import (
     17     _cy_wrapper_centered_discrepancy,
     18     _cy_wrapper_wrap_around_discrepancy,

ModuleNotFoundError: No module named 'scipy.stats._qmc_cy'

```

I suppose it doesn't matter since it has nothing to do with this PR.