@tupui 2021-04-10
How are you building and which platform are you on?

I just checked out your branch, built and I could import `from scipy.stats import invgauss`.
Are you certain you have pulled everything from master?