@zoj613 2021-04-10
> How are you building and which platform are you on?
> 
> I just checked out your branch, built and I could import `from scipy.stats import invgauss`.
> Are you certain you have pulled everything from master?

I think I did pull everything because I rebased using the upstream master branch and it says this branch is up to date on the forked repo. I'm on Manjaro Linux 64bit. I guess it's not too big of a deal because the tests run, just unable to import on ipython locally. Another issue is the faliure to build docs. I noticed this is the case for other open PR's as well.