@zoj613 2021-04-10
> Do you actually have the files `scipy/stats/_gmc_cy.pyx` and the `cxx, pyi` ones?
yes, a shell search returns:
```
$ ls scipy/stats/_qmc
_qmc_cy.cxx  _qmc_cy.pyi  _qmc_cy.pyx  _qmc.py      _qmc.pyi  
```