@tupui 2021-04-10
But then I would suspect that you have something in your PYTHONPATH or in the way you installed which prevents you from importing correctly. Did you follow the ubuntu guidelines from the devdoc? (https://scipy.github.io/devdocs/dev/contributor/quickstart_ubuntu.html#quickstart-ubuntu)
Especially the part about `conda develop .`