@zoj613 2021-04-10
> But then I would suspect that you have something in your PYTHONPATH or in the way you installed which prevents you from importing correctly. Because the tests are passing and I cannot reproduce with your branch on my side.
> 
> Did you follow the ubuntu guidelines from the devdoc? ([scipy.github.io/devdocs/dev/contributor/quickstart_ubuntu.html#quickstart-ubuntu](https://scipy.github.io/devdocs/dev/contributor/quickstart_ubuntu.html#quickstart-ubuntu))
> Especially the part about `conda develop .`

I read the page sometime ago but it did not make sense for me try reproduce those steps since I use `pyenv` to manage my python environments and dont have conda installed. It works fine to just create an environment for scipy using `pyenv` and then run `python setup.py build`. I didnt have issues this way until I updated the branch today.