@tupui 2021-04-10
> > Do you actually have the files `scipy/stats/_gmc_cy.pyx` and the `cxx, pyi` ones?
> 
> yes, a shell search returns:
> 
> ```
> $ ls scipy/stats/_qmc
> _qmc_cy.cxx  _qmc_cy.pyi  _qmc_cy.pyx  _qmc.py      _qmc.pyi  
> ```

Actually from this, I don't see the compiled module. After building SciPy, you should have a `_qmc_cy.cpython-....so`. So maybe it does have something to do with your platform. Check if in the `stats` folder you have other `.so`. You could try to just run the `stats/setup.py` and inspect the log to see if there is any build error.