@zoj613 2021-04-10
Update on performance based on the latest commit:

Master:
```python
In [2]: p = [0.000001, 0.00001, 0.0001, 0.001, 0.01, 0.1, 0.5, 0.9, 0.99,
   ...:              0.999, 0.9999, 0.99999, 0.999999]

In [3]: %timeit invgauss.ppf(p, mu=1)
51.7 ms ± 1.45 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [4]: %timeit invgauss.isf(p, mu=1)
51.5 ms ± 1.53 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

This PR:
```python
n [3]: p = [0.000001, 0.00001, 0.0001, 0.001, 0.01, 0.1, 0.5, 0.9, 0.99,
   ...:              0.999, 0.9999, 0.99999, 0.999999]

In [4]: %timeit invgauss.ppf(p, mu=1)
9.62 ms ± 96.9 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

In [5]: %timeit invgauss.isf(p, mu=1)
8.63 ms ± 42.2 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)
```